+++ apps/builder/src/modules/Style/StyleManager/modes/ManagerModeAdvanced.js
@@ -9,15 +9,13 @@ import CodeMirror from '@plitzi/plitzi-ui-components/CodeMirror';
 
 // Monorepo
 import EventBridgeTypes from '@plitzi/sdk-event-bridge/EventBridgeHelper';
+import { cssToSelectors, StyleConstants, getReadOnlyRangesFromContent } from '@plitzi/sdk-style/StyleHelper';
 
 // Alias
 import AppContext from '@pmodules/App/AppContext';
 import BuilderContext from '@pmodules/Builder/BuilderContext';
 import useNetwork from '@pmodules/Network/hooks/useNetwork';
 import NetworkContext from '@pmodules/Network/NetworkContext';
-
-// Relatives
-import { cssToSelectors, StyleConstants, getReadOnlyRangesFromContent } from '../../StyleHelper';
 
 const selectorsDefault = [];