+++ apps/builder/src/modules/Style/hooks/useStyleInherit.js
@@ -2,13 +2,13 @@
 import { useContext, useMemo } from 'react';
 import { ComponentContext } from '@plitzi/plitzi-sdk';
 
+// Monorepo
+import { calculateInheriting } from '@plitzi/sdk-style/StyleHelper';
+
 // Alias
 import AppContext from '@pmodules/App/AppContext';
 import BuilderStyleContext from '@pmodules/Builder/contexts/BuilderStyleContext';
 import BuilderSchemaContext from '@pmodules/Builder/contexts/BuilderSchemaContext';
-
-// Relatives
-import { calculateInheriting } from '../StyleHelper';
 
 const useStyleInherit = props => {
   const { element, selector, styleSelector } = props;
@@ -21,10 +21,22 @@ const useStyleInherit = props => {
     style: { platform }
   } = useContext(BuilderStyleContext);
 
-  const inheritData = useMemo(
-    () => calculateInheriting(element, flat, platform, displayMode, styleSelector, componentDefinitions, [selector]),
-    [element, flat, displayMode, styleSelector, selector, componentDefinitions]
-  );
+  const inheritData = useMemo(() => {
+    const selectorsToSkip = [];
+    if (!selector.includes(':')) {
+      selectorsToSkip.push(selector);
+    }
+
+    return calculateInheriting(
+      element,
+      flat,
+      platform,
+      displayMode,
+      styleSelector,
+      componentDefinitions,
+      selectorsToSkip
+    );
+  }, [element, flat, displayMode, styleSelector, selector, componentDefinitions]);
 
   return inheritData;
 };