+++ apps/builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @plitzi/plitzi-builder
 
+## 0.20.7
+
+### Patch Changes
+
+- v0.20.7
+- Updated dependencies
+  - @plitzi/sdk-shared@0.20.7
+  - @plitzi/plitzi-sdk@0.20.7
+  - @plitzi/sdk-event-bridge@0.20.7
+  - @plitzi/sdk-interactions@0.20.7
+  - @plitzi/sdk-plugins@0.20.7
+
 ## 0.20.6
 
 ### Patch Changes