+++ packages/sdk-interactions/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @plitzi/sdk-interactions
 
+## 0.20.7
+
+### Patch Changes
+
+- v0.20.7
+- Updated dependencies
+  - @plitzi/sdk-shared@0.20.7
+
 ## 0.20.6
 
 ### Patch Changes