+++ packages/sdk-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @plitzi/sdk-shared
 
+## 0.20.7
+
+### Patch Changes
+
+- v0.20.7
+
 ## 0.20.6
 
 ### Patch Changes